change GeoJSON context to zustand state management
