+++ components/Maps/GeoJSONLayer/index.tsx
@@ -1,13 +1,14 @@
 import { GeoJSON } from "./GeoJSON";
 import { LatLngBounds } from "leaflet";
 import { useMap, FeatureGroup } from "react-leaflet";
-import { useContext, useEffect, useRef } from "react";
-import { GeoJSONContext } from "@context/GeoJSONContext";
+import { useEffect, useRef } from "react";
+import { useGeoJSONStore } from "lib/store/geoJSONStore";
 
 export const GeoJSONLayer = () => {
   const map = useMap()
   const geojsonRef = useRef<L.FeatureGroup>(null)
-  const { geoJSON, mapKey } = useContext(GeoJSONContext)
+  const mapKey = useGeoJSONStore(state => state.mapKey)
+  const geoJSON = useGeoJSONStore(state => state.geoJSON)
 
   useEffect(() => {
     geoJSON && map.fitBounds(geojsonRef.current?.getBounds() as LatLngBounds)
