+++ lib/provider/AppProvider.tsx
@@ -1,15 +1,12 @@
 import { MapProvider } from './MapProvider';
-import { DataProvider } from './DataProvider';
 import { MantineProvider } from './MantineProvider';
 
 const AppProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => (
-  <DataProvider>
-    <MantineProvider>
-      <MapProvider>
-        {children}
-      </MapProvider>
-    </MantineProvider>
-  </DataProvider>
+  <MantineProvider>
+    <MapProvider>
+      {children}
+    </MapProvider>
+  </MantineProvider>
 )
 
 export default AppProvider
