--- lib/provider/DataProvider.tsx
@@ -1,7 +1,0 @@
-import { GeoJSONProvider } from '@context/GeoJSONContext';
-
-export const DataProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => (
-  <GeoJSONProvider>
-    {children}
-  </GeoJSONProvider>
-)
